prevent file overwrite, don't write deprecated sdsslink files

+++ aspectui_csv_list_input.py
@@ -102,16 +102,19 @@ for row in mapping_data_file:
         
         #spectra meta data
         spec_meta_data_file_path = ''.join((spec_meta_data_directory, '/', str(som_x), '-', str(som_y), '.json'))
-        with open(spec_meta_data_file_path, 'w') as spec_meta_data_file:
-            json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid), "sdsslink": link, "som_x": int(som_x), "som_y": int(som_y)}, spec_meta_data_file)
+        if not os.path.exists(spec_meta_data_file_path):
+            with open(spec_meta_data_file_path, 'w') as spec_meta_data_file:
+                json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid), "sdsslink": link, "som_x": int(som_x), "som_y": int(som_y)}, spec_meta_data_file)
         
         #write idmapping som_x, som_y -> mjd, plateid, fiberid and vice versa
         idmapping_file_path = ''.join((idmapping_directory, '/', str(som_x), '-', str(som_y), '.json'))
-        with open(idmapping_file_path, 'w') as idmapping_file:
-            json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid) }, idmapping_file)
+        if not os.path.exists(idmapping_file_path):
+            with open(idmapping_file_path, 'w') as idmapping_file:
+                json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid) }, idmapping_file)
         idmapping_file_path = ''.join((idmapping_directory, '/', str(mjd), '-', str(plateid), '-', str(fiberid), '.json'))
-        with open(idmapping_file_path, 'w') as idmapping_file:
-            json.dump({"som_x": int(som_x), "som_y":int(som_y)}, idmapping_file)            
+        if not os.path.exists(idmapping_file_path):
+            with open(idmapping_file_path, 'w') as idmapping_file:
+                json.dump({"som_x": int(som_x), "som_y":int(som_y)}, idmapping_file)            
         
         
         #make empty.png in highest zoom level if it not exists
